+ Working ConflictTest for a simple file
We have to take another look at the conflict handling. Handling the
conflict in ConflictState.execute() is probably to late. It has to be
done as soon as a conflicting event is catched (if it is a remote event)
since otherwise H2H just appends the content of the new file to the old,
since the old one is not renamed yet.