This commit contains a first prototype of the selective synchronization. It is more like a interactive test with dummy data than actual integration with the system. The reason behind is that all the objects and data structures needed for the selective sync to work are only present if the whole application has been initialized (i.e. FileManager to get files from the network, FileEventManager to get the local view of the files). As it is impractical to test the selective sync GUI only by initializing the whole application, several actions had to be taken.
First of all, interfaces IUserConfig, IFileEventManager, and
IPeerBoxFileManager are used as types of the injected parameters in the
constructor of Synchronization.java. This allows to start the selective
sync as well using Dummy-Managers and achieve minor isolation from the
rest of the app.

However, this implies changes in the real code: Switching to the real
data is only possible by changing the type of the parameters of
Synchronization.java's constructor. This is bad, but better then no
isolation.

Further more, a private constructor of
org.hive2hive.core.processes.files.list.FileNode has been changed to
public for testing purposes. This is bad as well, but the only chance to
generate dummy data.
 

The dummy data types and the SychronizationTestUtils are in the
productive source tree and NOT in the test tree as it has not been
recognized in the test tree. This has to be solved as well to separate
testing code from productive code.

For more information concerning the implementation details of the
Selective sync, contact me, or read the (soon available) documentation.